// types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      tech_stack: {
        Row: {
          category: string
          created_at: string
          icon_slug: string
          id: string
          name: string
        }
        Insert: {
          category: string
          created_at?: string
          icon_slug: string
          id?: string
          name: string
        }
        Update: {
          category?: string
          created_at?: string
          icon_slug?: string
          id?: string
          name?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}